parsePDFs02/requirements.txt
pypdf>=4.2.0
tabula-py>=2.9.3
openpyxl>=3.1.2
jpype1>=1.5.0
pymupdf>=1.24.4